test: add test for class with nested properties
This test currently fails to compile with the following error:
> Task :shared:compileTestKotlinJvm FAILED
e: file:///<...>/mockative/shared/build/generated/ksp/jvm/jvmTest/kotlin/io/github/Level1Mock.kt:22:12 Unresolved reference: Level2Mock

No mock class is generated for a mockable classes that is not used directly as a mock but indirectly as a property of another mockable class that in turn is used as a property of the mocked class. This leads to the compiler error 'Unresolved reference'.

The issue was likely introduced with 1241ddf5ec3896ac14918bec4b2962d1f9ca9028 removing the recursion in List<ProcessableType>.flatten().